Create cows_bulls_game
'cows and bulls' is a game where the user guesses the positions and values of 4 integers. For example 1-2-3-4. If the user guesses 1-3-2-4, the program returns: "2 cows, 2 bulls". A 'cow' is a correct number in the correct position (1 and 4) and a 'bull' is a correct number in the wrong position(3 and 2). When the user guesses the correct order of integers the program exits.